allow creating arbitrary resources on namespaces (#91)
* add resource quota to namespaces

* allow creation of arbitrary manifests during namespace creation

allows for ResourceQuotas, NetworkPolicies, etc. to be created prior to job submission
so limits can be enforced.

done with server-side apply, so we don't need to understand what kind of resources they are

Currently still a single limit, not customizable per user.

* test namespace manifests

moves derived settings to cached properties on Settings for easier reload / modification in tests

* update linter action